+++ Source/WebGPU/WebGPU/APIConversions.h
@@ -36,6 +36,7 @@
 #import "Device.h"
 #import "Instance.h"
 #import "PipelineLayout.h"
+#import "PresentationContext.h"
 #import "QuerySet.h"
 #import "Queue.h"
 #import "RenderBundle.h"
@@ -44,8 +45,6 @@
 #import "RenderPipeline.h"
 #import "Sampler.h"
 #import "ShaderModule.h"
-#import "Surface.h"
-#import "SwapChain.h"
 #import "Texture.h"
 #import "TextureView.h"
 #import <wtf/text/WTFString.h>
@@ -149,14 +148,14 @@ inline ShaderModule& fromAPI(WGPUShaderModule shaderModule)
     return static_cast<ShaderModule&>(*shaderModule);
 }
 
-inline Surface& fromAPI(WGPUSurface surface)
+inline PresentationContext& fromAPI(WGPUSurface surface)
 {
-    return static_cast<Surface&>(*surface);
+    return static_cast<PresentationContext&>(*surface);
 }
 
-inline SwapChain& fromAPI(WGPUSwapChain swapChain)
+inline PresentationContext& fromAPI(WGPUSwapChain swapChain)
 {
-    return static_cast<SwapChain&>(*swapChain);
+    return static_cast<PresentationContext&>(*swapChain);
 }
 
 inline Texture& fromAPI(WGPUTexture texture)
