+++ Source/WebGPU/WebGPU/Instance.mm
@@ -29,7 +29,7 @@
 #import "APIConversions.h"
 #import "Adapter.h"
 #import "HardwareCapabilities.h"
-#import "Surface.h"
+#import "PresentationContext.h"
 #import <cstring>
 #import <wtf/BlockPtr.h>
 #import <wtf/StdLibExtras.h>
@@ -65,9 +65,9 @@ Instance::Instance()
 
 Instance::~Instance() = default;
 
-Ref<Surface> Instance::createSurface(const WGPUSurfaceDescriptor& descriptor)
+Ref<PresentationContext> Instance::createSurface(const WGPUSurfaceDescriptor& descriptor)
 {
-    return Surface::create(descriptor);
+    return PresentationContext::create(descriptor);
 }
 
 void Instance::scheduleWork(WorkItem&& workItem)
@@ -462,6 +462,11 @@ WGPUProc wgpuGetProcAddress(WGPUDevice, const char* procName)
     if (!strcmp(procName, "wgpuTextureDestroy"))
         return reinterpret_cast<WGPUProc>(&wgpuTextureDestroy);
     return nullptr;
+}
+
+WGPUSurface wgpuInstanceCreateSurface(WGPUInstance instance, const WGPUSurfaceDescriptor* descriptor)
+{
+    return WebGPU::releaseToAPI(WebGPU::fromAPI(instance).createSurface(*descriptor));
 }
 
 void wgpuInstanceProcessEvents(WGPUInstance instance)
