--- Source/WebGPU/WebGPU/SwapChain.mm
@@ -1,72 +1,0 @@
-/*
- * Copyright (c) 2021-2022 Apple Inc. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY APPLE INC. ``AS IS'' AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL APPLE INC. OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
- * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
- * OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-#import "config.h"
-#import "SwapChain.h"
-
-#import "APIConversions.h"
-#import "Device.h"
-#import "TextureView.h"
-#import <WebGPU/WebGPUExt.h>
-
-namespace WebGPU {
-
-Ref<SwapChain> Device::createSwapChain(WGPUSurface surface, const WGPUSwapChainDescriptor& descriptor)
-{
-    return SwapChain::create(WTFMove(surface), descriptor);
-}
-
-SwapChain::SwapChain(WGPUSurface surface, const WGPUSwapChainDescriptor& descriptor)
-    : m_surface(WebGPU::fromAPI(surface))
-    , m_drawingBuffer(m_surface->drawingBuffer())
-{
-    UNUSED_PARAM(descriptor);
-}
-
-SwapChain::~SwapChain() = default;
-
-void SwapChain::present()
-{
-    m_drawingBuffer = m_surface->nextDrawable();
-}
-
-} // namespace WebGPU
-
-#pragma mark WGPU Stubs
-
-void wgpuSwapChainRelease(WGPUSwapChain swapChain)
-{
-    WebGPU::fromAPI(swapChain).deref();
-}
-
-WGPUTextureView wgpuSwapChainGetCurrentTextureView(WGPUSwapChain)
-{
-    return nullptr;
-}
-
-void wgpuSwapChainPresent(WGPUSwapChain swapChain)
-{
-    WebGPU::fromAPI(swapChain).present();
-}
